backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
numpy==1.26.4
librosa==0.10.2.post1
soundfile==0.12.1
audioread==3.0.1
ffmpeg-python==0.2.0
python-dotenv==1.0.1
openai==1.46.0
google-generativeai==0.7.2
TTS==0.22.0
